fix(Feed): 개별 피드 컴포넌트 세분화
