+++ src/pages/IndividualFeed.jsx
@@ -1,29 +1,15 @@
-import React, { useCallback, useEffect, useState } from "react";
-import backgroundImage from "../assets/images/IndividualFeed-BackgroundImage.png";
-import logo from "../assets/images/logo.png";
-import message from "../assets/images/Messages.png";
-import FeedCard from "../components/FeedCard";
+import React, { useEffect, useState } from "react";
 import AddQuestion from "../components/AddQuestion";
-import {
-  Wrapper,
-  Logo,
-  Profile,
-  BodyWrapper,
-  ProfilePlaceholder,
-  EmptyIcon,
-} from "../styles/individualFeedStyle";
+import { Wrapper } from "../styles/individualFeedStyle";
 import { useSubjectInfo } from "../hooks/useSubjectInfo";
 import { useIndividualQuestions } from "../hooks/useIndividualQuestions";
-import FeedCardPlaceholder from "../components/FeedCardPlaceholder";
 import { RotatingAnimation } from "../styles/rotatingAnimation";
 import { useScroll } from "../hooks/useScroll";
-import IconBox from "../components/IconBox";
 import Modal from "../components/Modal/Modal";
-import emptyIcon from "../assets/images/NoQuestion.svg";
-import { Link, useNavigate } from "react-router-dom";
+import FeedHeader from "../components/FeedHeader";
+import FeedBody from "../components/FeedBody";
 
 const IndividualFeed = () => {
-  const navigate = useNavigate();
   const [offset, setOffset] = useState(0);
   const LIMIT = 10;
   const { userInfo } = useSubjectInfo();
@@ -46,10 +32,6 @@ const IndividualFeed = () => {
     setSend(false);
   };
 
-  const navigateToMain = () => {
-    navigate("/");
-  };
-
   // 스켈리톤 ui를 위한 상태 변경
   useEffect(() => {
     if (userInfo) {
@@ -61,49 +43,14 @@ const IndividualFeed = () => {
 
   return (
     <Wrapper>
-      {/* 배경사진, 사용자 이름은 userInfo에서 가져온다 */}
-      <img src={backgroundImage} alt="배경사진" />
-      <Logo src={logo} alt="로고" onClick={navigateToMain} />
-      {loading ? (
-        <ProfilePlaceholder />
-      ) : (
-        <Profile src={userInfo?.imageSource} />
-      )}
-      <span className="profileName">{userInfo?.name}</span>
+      <FeedHeader loading={loading} userInfo={userInfo} />
 
-      {/* 아이콘 컴포넌트 */}
-      <IconBox />
-
-      {/* 질문을 보여주는 부분 */}
-      <BodyWrapper count={count}>
-        <div className="questionNum">
-          <img src={message} alt="질문 아이콘" />
-          <span>
-            {count === 0
-              ? "아직 질문이 없습니다."
-              : `${count}개의 질문이 있습니다.`}
-          </span>
-        </div>
-
-        {loading ? (
-          <>
-            {Array.from({ length: 5 }).map((_, index) => (
-              <FeedCardPlaceholder key={index} />
-            ))}
-          </>
-        ) : questionInfo?.length > 0 ? (
-          questionInfo.map((question, index) => (
-            <FeedCard
-              question={question}
-              key={index}
-              userName={userInfo.name}
-              profileImage={userInfo.imageSource}
-            />
-          ))
-        ) : (
-          <EmptyIcon src={emptyIcon} alt="질문 없을 때 이미지" />
-        )}
-      </BodyWrapper>
+      <FeedBody
+        count={count}
+        loading={loading}
+        questionInfo={questionInfo}
+        userInfo={userInfo}
+      />
 
       {moreData && questionInfo.length < count && <RotatingAnimation />}
